+++ packages/perps-sdk-provider-hyperliquid/src/accountConfig.ts
@@ -60,6 +60,7 @@ function projectHyperliquidDescriptor(
     case ActionType.UPDATE_ASSET_COLLATERAL:
     case ActionType.REGISTER_API_KEY:
     case ActionType.APPROVE_READ_ONLY_TOKEN:
+    case ActionType.SIWE_LOGIN:
     case ActionType.DEPOSIT:
     case ActionType.META_VOTE:
     case ActionType.META_ACCEPT_TERMS: